Refactor - moved files
Moved test files into better named packages, in a better-structured directory structure.
